backend_backup_20250830_163812/requirements-minimal.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.11.3
pydantic==1.10.7
SQLAlchemy==2.0.16
aiofiles==23.1.0
pytest==7.4.0
httpx==0.24.1
python-multipart==0.0.6
python-dotenv==1.0.0
